requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
python-dotenv==1.1.0
psycopg2-binary==2.9.10
pydantic==2.11.3
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tensorflow==2.20.0
numpy==2.3.3
